Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ int i;
 long int k;
 long long int j;
 float x;
-printf("Size of a char: %ld byte(s)\n", sizeof(s));
-printf("Size of an int: %ld byte(s)\n", sizeof(i));
-printf("Size of a long int : %ld byte(s)\n", sizeof(k));
-printf("Size of a long long int : %ld byte(s)\n", sizeof(j));
-printf("Size of a float: %ld byte(s)\n", sizeof(x));
+printf("Size of a char: %ld byte(s)\n",(unsigned long)sizeof(s));
+printf("Size of an int: %ld byte(s)\n",(unsigned long) sizeof(i));
+printf("Size of a long int : %ld byte(s)\n", (unsigned long)sizeof(k));
+printf("Size of a long long int : %ld byte(s)\n", (unsigned long)sizeof(j));
+printf("Size of a float: %ld byte(s)\n", (unsigned long)sizeof(x));
 return (0);
 }
